Moved code into MultiField module
